requirements.txt
Flask==2.1.0
Flask-PyMongo==2.3.0
Flask-Cors==3.0.10
pika==1.3.2